requirements.txt
streamlit
openai
python-dotenv
pyyaml
requests
pandas
